tdd_ai_agent/pyproject.toml
[project]
name = "tdd-ai-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "gengar", email = "mps.bigg@gmail.com" }
]
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.116.1",
    "langchain>=0.3.27",
    "langchain-community>=0.3.27",
    "langchain-experimental>=0.3.4",
    "langchain-openai>=0.3.29",
    "python-dotenv>=1.1.1",
    "python-multipart>=0.0.20",
    "uvicorn[standard]>=0.35.0",
]

[project.scripts]
tdd-ai-agent = "tdd_ai_agent:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
